extract example code to separate file
